Move referral codes above overview, restore chart size FEDEV-4162
Per design feedback on the ticket: shrinking the charts and packing them
into one row made them hard to read and misaligned the data points.
Instead the Referral codes block moves above the Overview block so
affiliates see their codes first, and the charts go back to their
original 2-up layout at 250px.

Reverts the chart height, the four-in-a-row grid (incl. the 1500px
breakpoint tweak), the card header restructure and the ScaledText
extraction, which existed only to support the narrow cards.
